Hand Skeleton Detection
Working hand skeleton detection system.  Takes in a videostream from a webcam and outputs the x, y, and z coordinates of the center of the hand.  The window has been edited to provide an easy PR system to drive the robot simply by manuevering your hand across the screen.  Circles are used as reference points.
**NOTE: The thresholds change when the distance to the camera changes.  These MUST be tuned prior to usage.  For best results, place a piece of tape on the ground and have people stand on the tape so the distance to the camera is consistent.